New Feature: randomly showing GIF from an directory
It is now possible not only to select a single GIF file but a directory
instead of it. If this directory includes GIF files they are randomly
played each time the screensaver starts. There is also a configurable
timer from 1 to 30 minutes in the options sections to define an
interval when the screensaver automatically changes the GIF randomly
during running.