enable a second screensaver instance when background mode is running
This is a problem or intension of implementation of the
ScreenSaverEngine of Apple. When running it with parameter
"-background" a second instance of a screensaver will not start(e.g.
after inactivity of user or moving mouse to an active corner).

I found out that calling the screensaverenginge with parameter
"-idlecheck" a second time after the background mode is already running
will stop this lock and enable a second screensaver instance afterwards.