Fix for issue that glview is not working(stay white) after return from sleep
AnimatedGif now register for event that is fired after return from
sleep.  It than simply kill screensaverengine. If it was in background
mode the launchd should restart the screensaver in background mode
again. It is a bit hard but works until I found a cleaner way to
restore the glview that stays white.