Issues #18 an #19
#18: The confusing GUI element for file / directory mode is removed.
Therefore when opening the file dialog not the last file or directory
is open, but one or two levels above, so that user see more of the path
structure with makes use of file OSX dialog more intuitive in my
opinion.

#19: The slider for selecting fps is now from 1 to 60 fps extended (old
was 1 to 30 fps). Remember if you disable manual selecting of fps it
uses the value of the GIF file, what ever this is (maybe 70 fps).