Add Workaround for #28
This is a workaround for issue #28 .
Since background-mode of ScreenSaverEngine is broken since 10.13
(screensaver is now always in an own space in foreground) AnimatedGif
uses the window-mode of ScreenSaverEngine instead and change the window
to a background window afterwards.

This is an initial commit. I need to test this Workaround more for
possible side effects before an official release.