Implementing Metal Support - part 5
Implementing the most of the missing render functions (coordinate system, scaling, draw picture and draw text) for Metal and so the most screensaver function working now also with Metal.
Anyway, there are still some TODOs in code (e.g. there is no double buffering at the moment or the Alpha Blending is disabled, because it was not working as expected with all test gif files).

-> Metal code still disabled by not defining USE_METAL