Finish Metal support #43
Finish implementing Metal support according to enhancement #43 .

Since it is the biggest internal change since implementing OpenGL the version is increased from 1.4.0 to 1.5.0 (skipping the intermete version 1.4.1 used during implementation).

In the commit included is a ZIP with the compiled screensaver. But before I make an official GIT release of 1.5.0 I need to do more tests on other systems I have.