Add GIF grammar file for "Synalyze It!"
Add a GIF grammar file for "Synalyze It!" (and probably Hexinator too?) with makes it a bit more easy to analyse strange GIF files.

I created this GIF grammar file all by myself since the default GIF grammar from "Synalyze It!" was too basic and did not show more than the header and the LogicalScreenDescriptor.

There are still a few things that probably not completely right in this new GIF grammar file especially the structure different types of Application Extensions are not well documented and a kind of a mystery.