Change signing
Change signing from a self created certificat to a free "Mac Developer" certificat from Apple root signed.

This is still no "Developer ID" account that coste 99$ a year and I still don't get the notarizing of the app working with this certificat. It seams also  notarizing of the app is only working with a paid "Developer ID".

I also update with this free "Mac Developer" certificat the compiled screensaver in the zip file from the latest 1.5.3 release .