#74: Workaround for ScreenSaverDefaults not updated in Sonoma
ScreenSaverDefaults are not updated in macOS Sonoma when changed in System Preferences because an old instance of legacyScreenSaver with older values is still running in background and re-used when activating the screen saver again instead closing of legacyScreenSaver and starting an new instance with new values from ScreenSaverDefaults.
Adding a notification handler to com.apple.screensaver.willstop with exit(0) will now quit this old background instance when the screensaver is stoped used instead of keeping it running and reuing it later.